Replaced source:/gplates/trunk using source:/gplates/branches/shapefile-2007-nov-05#2193 incorporating log:/gplates/branches/shapefile-2007-nov-05#2041:2193 ([log:/gplates/trunk#2192:2193 No changes on trunk] since last update from trunk
This is "[wiki:MergingInSubversion#Replacingonebranchwithanother replacing one branch with another]". The command used was:
{{{
svn merge http://svn.gplates.org/gplates/trunk@2193 http://svn.gplates.org/gplates/branches/shapefile-2007-nov-05@2193
}}}
from inside the root directory of a working copy of the trunk.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@2194 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/gui/Globe.cc
@@ -216,7 +216,10 @@ GPlatesGui::Globe::Paint()
 		           axis.x().dval(), axis.y().dval(), axis.z().dval());
 
 		// Set the sphere's colour.
-		glColor3fv(GPlatesGui::Colour(0.35, 0.35, 0.35));
+		glColor3fv(GPlatesGui::Colour(
+				static_cast<GLfloat>(0.35), 
+				static_cast<GLfloat>(0.35), 
+				static_cast<GLfloat>(0.35)));
 		
 		/*
 		 * Draw sphere.
